Move query-string from flux/ to lib/
